this the excel file to analyst the data
this data was gain from the survey conducted within faculty of ICT in IIUM. the respondents are students of faculty of ICT in IIUM. the survey conducted by interviewing the each of the students. the students will answer the question by using my smartphone since the question is conducted by using google form. the students consist of local and international students